Move the DbInitializers to its own namespace
